+++ tests/testthat/test-desirability.R
@@ -23,7 +23,7 @@ test_that("basic desirability functions", {
       target(a, low = 1, target = 2, high = 3, scale_low = 2),
       category(b, categories = list(a = 0.1, b = 0.2, c = 0.3)),
       constrain(x, low = 1, high = 2),
-      .use_data = TRUE
+      .use_data = FALSE
     )
   expect_snapshot(
     d2@inputs