+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Export a helper function (`make_desirability_cols()`) to make it easier for 
   other packages to import and use desirability2. 
+  
+* The `.use_data` default for `desirability()` was changed to `TRUE`. 
 
 # desirability2 0.1.0